pyproject.toml
[tool.poetry]
name = "blueberry"
version = "0.1.0"
description = "Interface de supervision NIDS"
authors = ["rgd <michael@mic-rigaud.fr>"]

[tool.poetry.dependencies]
python = "^3.7.3"
invoke = "^1.6.0"
pyzmq = "^22.3.0"
python-telegram-bot = "^13.8.1"
peewee = "^3.14.8"
psutil = "^5.8.0"
requests = "^2.26.0"
graphviz = "^0.18.2"
maxminddb-geolite2 = "^2018.703"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
bandit = "^1.7.1"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
